// src/model/jobSummary.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * K-Culture Web API
 * 배우담(K-Culture Web) 플랫폼 API 명세서

## 개요
- 배우와 에이전시를 연결하는 AI 기반 캐스팅 플랫폼
- 시놉시스 기반 배우 추천, 프로젝트 관리, 포트폴리오 관리 기능 제공

## 인증
- JWT Bearer Token 기반 인증
- Access Token 유효기간: 1시간
- Refresh Token 유효기간: 7일

 * OpenAPI spec version: 1.0.0
 */
import type { JobCategory } from './jobCategory';
import type { JobSummaryGender } from './jobSummaryGender';
import type { JobStatus } from './jobStatus';

export interface JobSummary {
  id: string;
  category: JobCategory;
  /** 품앗이 여부 */
  isPumasi?: boolean;
  /**
   * 출연료 (원)
   * @nullable
   */
  price?: number | null;
  title: string;
  gender?: JobSummaryGender;
  /** 제작사/제작팀 */
  production?: string;
  /** 작품 제목 */
  workTitle?: string;
  status: JobStatus;
  /** 조회수 */
  views?: number;
  createdAt?: string;
}